refactor(api)!: changement gestion client/hotelier
Les entités `client` et `hotel_owner` ont grandement changé et hérite
leurs informations, désormais de la nouvelle entité `user`.

Cela a été fait dans le futur d'ajouter le système d'authentification.
Django n'autorise qu'à utiliser UNE SEULE entité pour représenter un
utilisateur à authentifier, ce sera `user`. Les 2 entités ont donc été
modifiées afin de correctement hériter de `user`.

Désormais, un client sera une entité pointant vers un utilisateur. De
même pour un hotelier qui sera une entité pointant vers un utilisateur
(avec une information en plus `company`).

Les résultats des endpoints ont un peu changé pour s'adapter :
{
	'id',
	'info': {
		'id',
		'password',
		'last_login',
		'mail',
		'name',
		'lastname',
		'is_active',
		'staff',
		'admin'
		},
	'company': # si utilisateur `hotel_owner`
}

De plus, toutes les routes ont été modifiées. Aucun changement sur leur
utilisation. Le début des endpoit ont été 'déplacées' vers les routes globales
`big_b_n_b\urls.py` afin que Django montre le début de chaque endpoint
si aucune route n'était pas précisé.

+++ back/client/views.py
@@ -1,27 +1,38 @@
-from rest_framework import status
+from rest_framework import status, viewsets
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
 
 from client.models import Client
 from client.serializers import ClientSerializer
+from django.contrib.auth import get_user_model
+
+from user.serializers import UserSerializer, UserSerializerValidator
+
+User = get_user_model()
 
 
 @api_view(['GET', 'POST'])
-def all_clients(request):
+def all_clients(request) -> Response:
     if request.method == 'GET':
-        clients = Client.objects.all()
+        clients = Client.objects.select_related('user').all()
         return Response(ClientSerializer(clients, many=True).data, status=status.HTTP_200_OK)
 
     if request.method == 'POST':
-        serializer = ClientSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(status=status.HTTP_201_CREATED)
-        return Response(status=status.HTTP_400_BAD_REQUEST, data=serializer.errors)
+        validator: UserSerializerValidator = UserSerializerValidator(data=request.data)
+        if not validator.is_valid():
+            return validator.get_current_response()
+
+        validator.create_user()
+        data = validator.get_final_data()
+        client_serializer = ClientSerializer(data=data)
+        if not client_serializer.is_valid():
+            return Response(data=client_serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        client_serializer.save()
+        return Response(status=status.HTTP_201_CREATED)
 
 
 @api_view(['GET', 'DELETE'])
-def client_details(request, client_id):
+def client_details(request, client_id) -> Response:
     try:
         client = Client.objects.all().get(pk=client_id)
     except Client.DoesNotExist:
@@ -31,5 +42,5 @@ def client_details(request, client_id):
         return Response(ClientSerializer(client).data, status=status.HTTP_200_OK)
 
     if request.method == 'DELETE':
-        client.delete()
-        return Response(status=status.HTTP_200_OK)
+        User.objects.filter(id=client.user_id).delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
